Group Presentation Python Notebook
What we have currently. We can remove the bottom code that does not run, and replace with the new plots for the growth for the top few industries per year, and then TJ's visualizations, as they kind of sum up our data very well. May have to match variables, but maybe not